+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "The travelling salesman problem".
 - Tests for "Murder in the village!".
 - Tests for "Dead men's shot".
+- Tests for "Darts".
 
 ## [1.7.0] - 2022-03-31
 ### Added